+++ app/airflow/dags/concepts_from_vocab_dag.py
@@ -2,22 +2,20 @@ from datetime import datetime, timedelta
 from airflow import DAG
 from airflow.operators.empty import EmptyOperator
 from airflow.operators.python import PythonOperator
-from libs.core_create_concepts import (
+from libs.core_concepts_creation import (
     find_standard_concepts,
     create_standard_concepts,
-    find_date_fields,
-    find_additional_fields,
-    find_concept_fields,
-    find_dest_table_and_person_field_id,
     find_sr_concept_id,
 )
 
+from libs.core_prep_rules_creation import (
+    find_dest_table_and_person_field_id,
+    find_date_fields,
+    find_concept_fields,
+    find_additional_fields,
+)
+
 from libs.core_rules_creation import (
-    create_person_rules,
-    create_dates_rules,
-    create_concepts_rules,
-    add_rules_to_temp_table,
-    temp_mapping_rules_table_creation,
     delete_mapping_rules,
     create_mapping_rules,
 )
@@ -112,41 +110,6 @@ delete_mapping_rules_task = PythonOperator(
     dag=dag,
 )
 
-temp_task_2 = PythonOperator(
-    task_id="temp_mapping_rules_table_creation",
-    python_callable=temp_mapping_rules_table_creation,
-    provide_context=True,
-    dag=dag,
-)
-
-temp_task = PythonOperator(
-    task_id="add_rules_to_temp_table",
-    python_callable=add_rules_to_temp_table,
-    provide_context=True,
-    dag=dag,
-)
-
-create_person_rules_task = PythonOperator(
-    task_id="create_person_rules",
-    python_callable=create_person_rules,
-    provide_context=True,
-    dag=dag,
-)
-
-create_dates_rules_task = PythonOperator(
-    task_id="create_dates_rules",
-    python_callable=create_dates_rules,
-    provide_context=True,
-    dag=dag,
-)
-
-create_concepts_rules_task = PythonOperator(
-    task_id="create_concepts_rules",
-    python_callable=create_concepts_rules,
-    provide_context=True,
-    dag=dag,
-)
-
 create_mapping_rules_task = PythonOperator(
     task_id="create_mapping_rules",
     python_callable=create_mapping_rules,
@@ -156,24 +119,6 @@ create_mapping_rules_task = PythonOperator(
 
 # End the workflow
 end = EmptyOperator(task_id="end", dag=dag)
-
-# (
-#     start
-#     >> find_std_concepts_task
-#     >> create_standard_concepts_task
-#     >> find_sr_concept_id_task
-#     >> find_dest_table_and_person_field_task
-#     >> find_date_fields_task
-#     >> find_concept_fields_task
-#     >> find_additional_fields_task
-#     >> delete_mapping_rules_task
-#     >> temp_task_2
-#     >> create_person_rules_task
-#     >> create_dates_rules_task
-#     >> create_concepts_rules_task
-#     >> temp_task
-#     >> end
-# )
 
 (
     start
